refactor: props to attrs
