+++ packages/eslint-plugin-lit-a11y/lib/rules/aria-attrs.js
